Add try_clone to sync::ClientExt.
`SyncClient` is no longer `Clone` because it needs to create
a new `reactor::Core` when cloning. Instead, `try_clone` is available
on the client extension trait previously named `sync::Connect`.
Both client traits `sync::Connect` and `future::Connect` are renamed
to reflect the possibility of more fns added to them in the future (or
present, in the case of `sync::Connet`).