Fix request leak in the server.
When a request was expired, the server was dropping the response without decrementing the in-flight count, so it wasn't shutting down properly. I've changed it so that the response is always sent back to the client, even if it has reason to believe the client has lost interest in the request.