Impl Stub for Serve types.
In an interesting twist, it turns out that all tarpc services can also
act as client stubs. This makes sense intuitively: a stub just sends a
request to a service and waits to receive a response; if you already
have a service, you can simply cut out the middleman.

I'm not really sure when this would be useful. Has science gone too far?
Maybe it could be used to more easily mock out backends in unit tests?