Fix CI workflow file (#466)
* Fix CI workflow file

The final step didn't include the Clippy and format tests. So even if
those failed, PRs could get merged anyway. This now depends on all
steps.

Also, the cancel-workflow-action is no longer necessary, as stated in
the README of the workflow. GHA has implemented this feature by now.

Also removes the `check` step, as the final test didn't depend on it, it
didn't use `-Dwarnings`, and this is also covered by the `clippy` step.

This should now fail CI in the PR, as the clippy check currently fails.

* Fix Clippy error

* Bump MSRV to 1.65.0